[9.x] Refactor paginators
